Workaround for incompatibility with avr-gcc 3.6.1
There is a bug with the arduino-builder's sketch preprocessing when used with the avr-gcc 3.6.1 toolchain that will be in the next Arduino AVR Boards release. This causes compilation of sketches that use this library's macros to fail:

PalatisSoftPWM_example:7:1: error: expected unqualified-id before ')' token

 SOFTPWM_DEFINE_PIN13_CHANNEL(0);  //Configure Arduino pin 13 as PWM channel 0

I noticed that the bug did not occur with the Palatis/arduino-softpwm library. I found this was because Palatis had used a namespace, which happens to stave off the IDE from mucking with the code. I didn't want to change the library's API so I used a an unnamed namespace.